+++ src/engine/raft_store_engine.h
@@ -155,14 +155,14 @@ class RaftStoreEngine : public Engine, public RaftControlAble {
    public:
     TxnReader(std::shared_ptr<RawEngine> engine) : raw_engine_(engine) {}
 
-    butil::Status TxnBatchGet(std::shared_ptr<Context> ctx, uint64_t start_ts, const std::vector<std::string>& keys,
+    butil::Status TxnBatchGet(std::shared_ptr<Context> ctx, int64_t start_ts, const std::vector<std::string>& keys,
                               std::vector<pb::common::KeyValue>& kvs,
                               pb::store::TxnResultInfo& txn_result_info) override;
-    butil::Status TxnScan(std::shared_ptr<Context> ctx, uint64_t start_ts, const pb::common::Range& range,
-                          uint64_t limit, bool key_only, bool is_reverse, pb::store::TxnResultInfo& txn_result_info,
+    butil::Status TxnScan(std::shared_ptr<Context> ctx, int64_t start_ts, const pb::common::Range& range, int64_t limit,
+                          bool key_only, bool is_reverse, pb::store::TxnResultInfo& txn_result_info,
                           std::vector<pb::common::KeyValue>& kvs, bool& has_more, std::string& end_key) override;
-    butil::Status TxnScanLock(std::shared_ptr<Context> ctx, uint64_t min_lock_ts, uint64_t max_lock_ts,
-                              const pb::common::Range& range, uint64_t limit,
+    butil::Status TxnScanLock(std::shared_ptr<Context> ctx, int64_t min_lock_ts, int64_t max_lock_ts,
+                              const pb::common::Range& range, int64_t limit,
                               std::vector<pb::store::LockInfo>& lock_infos) override;
 
    private:
@@ -181,9 +181,9 @@ class RaftStoreEngine : public Engine, public RaftControlAble {
                                          const std::vector<std::string>& keys) override;
     butil::Status TxnPrewrite(std::shared_ptr<Context> ctx, const std::vector<pb::store::Mutation>& mutations,
                               const std::string& primary_lock, int64_t start_ts, int64_t lock_ttl, int64_t txn_size,
-                              bool try_one_pc, int64_t max_commit_ts, const std::vector<int32_t>& pessimistic_checks,
-                              const std::map<int32_t, int64_t>& for_update_ts_checks,
-                              const std::map<int32_t, std::string>& lock_extra_datas) override;
+                              bool try_one_pc, int64_t max_commit_ts, const std::vector<int64_t>& pessimistic_checks,
+                              const std::map<int64_t, int64_t>& for_update_ts_checks,
+                              const std::map<int64_t, std::string>& lock_extra_datas) override;
     butil::Status TxnCommit(std::shared_ptr<Context> ctx, int64_t start_ts, int64_t commit_ts,
                             const std::vector<std::string>& keys) override;
     butil::Status TxnCheckTxnStatus(std::shared_ptr<Context> ctx, const std::string& primary_key, int64_t lock_ts,
